All tests updated, dependencies clean up
