+++ Solana/bridge-token/workflows/src/integrationTest/kotlin/net/corda/samples/solana/bridge/token/DevNetBridgeTokenDemo.kt
@@ -5,13 +5,14 @@ import net.corda.testing.driver.DriverParameters
 import net.corda.testing.driver.driver
 import net.corda.testing.node.NotarySpec
 import org.junit.jupiter.api.Test
+import org.slf4j.LoggerFactory
 
-// Shareholder: https://solscan.io/account/3Wuk6fKtqCzMppikC1S58vK3J5ZbqnbcZXkDhJUHCom7?cluster=devnet#portfolio
-// Other Shareholder https://solscan.io/account/AoDHzQwk7s6crxMcC1nptRVHAhd1LASEWbQmAQjCeBKj?cluster=devnet#portfolio
 class DevNetBridgeTokenDemo : DevNetBridgeTokenTest() {
 
+    private val log = LoggerFactory.getLogger(DevNetBridgeTokenDemo::class.java)
+
     @Test
-    fun `briding token demo`() = driver(
+    fun `dev net bridge token demo`() = driver(
         DriverParameters(
             isDebug = false,
             inMemoryDB = false,
@@ -25,8 +26,11 @@ class DevNetBridgeTokenDemo : DevNetBridgeTokenTest() {
             waitForAllNodesToFinish = true
         )
     ) {
-        log.info("\nStarting bridging test using Solana validator via $solanaRpcUrl...")
+        log.info("\nStarting bridge test using Solana validator via $solanaRpcUrl...")
         runtimeSetup()
-        log.info("\nBridging deployment is running, shut down Corda nodes externally to exit...")
+        log.info("\nBridge deployment is running, shut down Corda nodes externally to exit...")
+        log.info("\nShareholders wallets:")
+        log.info("\n Shareholder: https://solscan.io/account/3Wuk6fKtqCzMppikC1S58vK3J5ZbqnbcZXkDhJUHCom7?cluster=devnet#portfolio")
+        log.info("\n Other Shareholder https://solscan.io/account/AoDHzQwk7s6crxMcC1nptRVHAhd1LASEWbQmAQjCeBKj?cluster=devnet#portfolio")
     }
 }
