+++ Solana/bridge-token/workflows/src/integrationTest/kotlin/net/corda/samples/solana/bridge/token/DevNetBridgeTokenTest.kt
@@ -8,50 +8,57 @@ import net.corda.testing.common.internal.testNetworkParameters
 import net.corda.testing.driver.DriverParameters
 import net.corda.testing.driver.driver
 import net.corda.testing.node.NotarySpec
+import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import org.slf4j.LoggerFactory
 import software.sava.core.accounts.PublicKey
 import java.math.BigDecimal
 import java.net.URI
 import java.nio.file.Path
 import java.nio.file.Paths
 
-open class DevNetBridgeTokenTest : BridgingTokenDriverTest() {
+open class DevNetBridgeTokenTest : TestBase() {
 
-    val solanaRpcUrl = "https://api.devnet.solana.com"
-    val solanaWssUrl = "ws://api.devnet.solana.com"
+    private val log = LoggerFactory.getLogger(DevNetBridgeTokenTest::class.java)
+
+    // A directory for Notary to store Corda participant key pairs for signing Solana transactions,
+    // intentionally these are located in a different directory than Corda Notary Program key pair
     protected val staticCustodiedKeysDir = "src/integrationTest/resources/custodiedKeys"
-    lateinit var solanaNotarySigner: FileSigner
+    protected lateinit var solanaNotarySigner: FileSigner
 
-    override fun getSolanaNotaryConfig(solanaNotarySigner: FileSigner) : Map<String, Any> {
-        return mapOf<String, Any>(
+    fun getSolanaNotaryConfig(solanaNotarySigner: FileSigner) = mapOf<String, Any>(
             "notary" to mapOf(
                 "validating" to false,
                 "solana" to mapOf(
                     "rpcUrl" to solanaRpcUrl,
-                    "websocketUrl" to solanaWssUrl,
+                    "websocketUrl" to solanaWebsocketUrl,
                     "notaryKeypairFile" to "${solanaNotarySigner.file}",
                     "custodiedKeysDir" to "${Path.of(staticCustodiedKeysDir).toAbsolutePath()}"
                 )
             )
         )
-    }
 
-    fun startTestValidator() {
+    @BeforeEach
+    fun setup() {
+        solanaRpcUrl = "https://api.devnet.solana.com"
+        solanaWebsocketUrl = "ws://api.devnet.solana.com"
         val notaryKeyPath =
-             Paths.get("../../Dev7chG99tLCAny3PNYmBdyhaKEVcZnSTp3p1mKVb5m5.json").toAbsolutePath().toString()
-        //solanaNotarySigner = FileSigner.read(Path.of(notaryKeyPath))
-        solanaClient = SolanaClient(URI.create(solanaRpcUrl), URI.create(solanaWssUrl)).apply { start() }
+            Paths.get("../../Dev7chG99tLCAny3PNYmBdyhaKEVcZnSTp3p1mKVb5m5.json").toAbsolutePath().toString()
+        solanaNotarySigner = FileSigner.read(Path.of(notaryKeyPath))
+        solanaClient = SolanaClient(URI.create(solanaRpcUrl), URI.create(solanaWebsocketUrl)).apply { start() }
         tokenManagement = TokenManagement(solanaClient)
         accountManagement = AccountManagement(solanaClient)
-    }
 
-    override fun setupAccounts() {
-        bridgeAuthoritySigner = FileSigner.read(Path.of("$staticCustodiedKeysDir/bridgeAuthority.json").toAbsolutePath())
+        bridgeAuthoritySigner =
+            FileSigner.read(Path.of("$staticCustodiedKeysDir/bridgeAuthority.json").toAbsolutePath())
         redemptionWalletForShareholder =
             FileSigner.read(Path.of("$staticCustodiedKeysDir/redemptionWalletForShareholder.json").toAbsolutePath())
         redemptionWalletForOtherShareholder =
-            FileSigner.read(Path.of("$staticCustodiedKeysDir/redemptionWalletForOtherShareholder.json").toAbsolutePath())
-        mintAuthoritySigner = FileSigner.read(Path.of("$staticCustodiedKeysDir/mintAuthoritySigner.json").toAbsolutePath())
+            FileSigner.read(
+                Path.of("$staticCustodiedKeysDir/redemptionWalletForOtherShareholder.json").toAbsolutePath()
+            )
+        mintAuthoritySigner =
+            FileSigner.read(Path.of("$staticCustodiedKeysDir/mintAuthoritySigner.json").toAbsolutePath())
         val otherKeysDir = "src/integrationTest/resources/other"
         shareholderWallet = FileSigner.read(Path.of("$otherKeysDir/shareholderWallet.json").toAbsolutePath())
         otherShareholderWallet = FileSigner.read(Path.of("$otherKeysDir/otherShareholderWallet.json").toAbsolutePath())
@@ -72,39 +79,39 @@ open class DevNetBridgeTokenTest : BridgingTokenDriverTest() {
         tokenMint2 = PublicKey.fromBase58Encoded("AhZNYSxWTVCbXMcZNJv27e7G38G7auFaqx4zCBBfWikc")
 
         log.info("  Creating Solana Token.")
-        log.info("  Shareholder Token Account: ${shareholderWallet.deriveATA().toBase58()}")
-        log.info("  Other Shareholder Token Account: ${otherShareholderWallet.deriveATA().toBase58()}")
-        log.info("  Shareholder Redemption Token Account: ${redemptionWalletForShareholder.deriveATA().toBase58()}")
+        log.info("  Shareholder Token Account: ${shareholderWallet.toATA().toBase58()}")
+        log.info("  Other Shareholder Token Account: ${otherShareholderWallet.toATA().toBase58()}")
+        log.info("  Shareholder Redemption Token Account: ${redemptionWalletForShareholder.toATA().toBase58()}")
         log.info(
             "  Other Shareholder Redemption Token Account: ${
-                redemptionWalletForOtherShareholder.deriveATA().toBase58()
+                redemptionWalletForOtherShareholder.toATA().toBase58()
             }"
         )
 
-        val shareholderBalance = solanaClient.getSolanaTokenBalance(shareholderWallet.deriveATA())
+        val shareholderBalance = solanaClient.getSolanaTokenBalance(shareholderWallet.toATA())
         if (shareholderBalance > BigDecimal.ZERO) {
             log.info("  Shareholder Token Account cleanup - burn existing $shareholderBalance tokens")
             tokenManagement.burn(
                 shareholderWallet,
                 tokenMint,
-                shareholderWallet.deriveATA(),
+                shareholderWallet.toATA(),
                 shareholderBalance.toLong()
             )
         }
-        val otherShareholderBalance = solanaClient.getSolanaTokenBalance(otherShareholderWallet.deriveATA())
+        val otherShareholderBalance = solanaClient.getSolanaTokenBalance(otherShareholderWallet.toATA())
         if (otherShareholderBalance > BigDecimal.ZERO) {
             log.info("  Other Shareholder Token Account cleanup - burn existing $otherShareholderBalance tokens")
             tokenManagement.burn(
                 otherShareholderWallet,
                 tokenMint,
-                otherShareholderWallet.deriveATA(),
+                otherShareholderWallet.toATA(),
                 otherShareholderBalance.toLong()
             )
         }
     }
 
     @Test
-    fun `dev net briding token test`() = driver(
+    fun `dev net bridge token test`() = driver(
         DriverParameters(
             isDebug = false,
             inMemoryDB = false,
